Adding UA Identifiers to Gimbal Device Attributes
Gimbal is adding the Urban Airship channel ID and named user ID to the
Gimbal Device Attributes.

The reason for this is to allow customers to link their UA Push
Messaging data to Gimbal Place Events in Gimbal Manager based on these
identifiers being related to the Gimbal Application Instance ID.

We also made some implementation changes to attempt to make the code
more readable and efficient. We did not make any public interface
changes.